finbuddy/backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
passlib[bcrypt]
openai
pydantic
pandas 